backend/requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
dj-database-url==0.5.0
Django==3.2.3
django-cors-headers==3.7.0
django-rest-swagger==2.2.0
django-secrets==1.0.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
gunicorn
drf-yasg==1.20.0
future==0.18.2
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.0
MarkupSafe==2.0.0
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.9
psycopg2-binary==2.8.6
PyJWT==2.1.0
pyparsing==2.4.7
PySocks==1.7.1
python-twitter==3.5
pytz==2021.1
repackage==0.7.3
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tweepy==3.10.0
uritemplate==3.0.1
urllib3==1.26.4
coreschema==0.0.4
django-heroku
apscheduler
